Refactor(jira): Move common classes into the common package.
