+++ src/main/java/com/synopsys/integration/alert/channel/jira/common/JiraMessageParser.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel.jira.cloud.util;
+package com.synopsys.integration.alert.channel.jira.common;
 
 import org.springframework.stereotype.Component;
