+++ src/main/java/com/synopsys/integration/alert/channel/jira/common/model/AlertJiraIssueProperties.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel.jira.cloud.model;
+package com.synopsys.integration.alert.channel.jira.common.model;
 
 import com.synopsys.integration.alert.common.rest.model.AlertSerializableModel;
